requirements.txt
pydantic>=2.7.0
jinja2>=3.1.4
openai>=1.35.0
tenacity>=8.3.0
python-dotenv>=1.0.1
pytest>=8.2.0
